server/requirements.txt
backoff==2.2.1 ; python_version >= "3.9" and python_version < "4.0"
bitsandbytes==0.41.3.post2 ; python_version >= "3.9" and python_version < "4.0"
boto3==1.34.4 ; python_version >= "3.9" and python_version < "4.0"
botocore==1.34.4 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
deprecated==1.2.14 ; python_version >= "3.9" and python_version < "4.0"
einops==0.6.1 ; python_version >= "3.9" and python_version < "4.0"
filelock==3.13.1 ; python_version >= "3.9" and python_version < "4.0"
fsspec==2023.10.0 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos==1.62.0 ; python_version >= "3.9" and python_version < "4.0"
grpc-interceptor==0.15.4 ; python_version >= "3.9" and python_version < "4.0"
grpcio-reflection==1.60.0 ; python_version >= "3.9" and python_version < "4.0"
grpcio-status==1.60.0 ; python_version >= "3.9" and python_version < "4.0"
grpcio==1.60.0 ; python_version >= "3.9" and python_version < "4.0"
hf-transfer==0.1.4 ; python_version >= "3.9" and python_version < "4.0"
huggingface-hub==0.19.4 ; python_version >= "3.9" and python_version < "4.0"
idna==3.6 ; python_version >= "3.9" and python_version < "4.0"
importlib-metadata==6.11.0 ; python_version >= "3.9" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.9" and python_version < "4.0"
loguru==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.26.2 ; python_version >= "3.9" and python_version < "4.0"
opentelemetry-api==1.21.0 ; python_version >= "3.9" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-common==1.21.0 ; python_version >= "3.9" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-grpc==1.21.0 ; python_version >= "3.9" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-http==1.21.0 ; python_version >= "3.9" and python_version < "4.0"
opentelemetry-exporter-otlp==1.21.0 ; python_version >= "3.9" and python_version < "4.0"
opentelemetry-instrumentation-grpc==0.42b0 ; python_version >= "3.9" and python_version < "4.0"
opentelemetry-instrumentation==0.42b0 ; python_version >= "3.9" and python_version < "4.0"
opentelemetry-proto==1.21.0 ; python_version >= "3.9" and python_version < "4.0"
opentelemetry-sdk==1.21.0 ; python_version >= "3.9" and python_version < "4.0"
opentelemetry-semantic-conventions==0.42b0 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.25.1 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
regex==2023.10.3 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
s3transfer==0.9.0 ; python_version >= "3.9" and python_version < "4.0"
safetensors==0.3.1 ; python_version >= "3.9" and python_version < "4.0"
sentencepiece==0.1.99 ; python_version >= "3.9" and python_version < "4.0"
setuptools==69.0.2 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
stanford-stk==0.0.6 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "linux"
tiktoken==0.5.2 ; python_version >= "3.9" and python_version < "4.0"
tokenizers==0.15.0 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.66.1 ; python_version >= "3.9" and python_version < "4.0"
transformers==4.36.0 ; python_version >= "3.9" and python_version < "4.0"
triton==2.1.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "linux"
typer==0.6.1 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.18 ; python_version >= "3.9" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
wrapt==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.17.0 ; python_version >= "3.9" and python_version < "4.0"
